requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
mypy==1.14.1
mypy-extensions==1.0.0
pydantic==2.10.5
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
tqdm==4.64.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0
httpx==0.24.0
sqlalchemy==2.0.25
pymysql==1.1.1
pymongo[srv]==4.3.3